Create separate bls12-381 module
